Move expensive InputSpecBuilders
Provide symbolic_expression in a separate template file.
