+++ src/core/io/src/4C_io_input_spec_builders.hpp
@@ -18,9 +18,9 @@
 #include "4C_io_input_types.hpp"
 #include "4C_io_value_parser.hpp"
 #include "4C_io_yaml.hpp"
+#include "4C_utils_compile_time_string.hpp"
 #include "4C_utils_enum.hpp"
 #include "4C_utils_string.hpp"
-#include "4C_utils_symbolic_expression.hpp"
 
 #include <algorithm>
 #include <functional>
@@ -980,20 +980,7 @@ namespace Core::IO
     };
 
     template <typename Number, Utils::CompileTimeString... variables>
-    struct SymbolicExpressionData
-    {
-      /**
-       * An optional description of the symbolic expression.
-       */
-      std::string description{};
-
-      /**
-       * An optional function to store the parsed symbolic expression. By default, the result is
-       * stored in an InputParameterContainer. See the in_struct() function for more details on how
-       * to store the symbolic expression in a struct.
-       */
-      StoreFunction<Utils::SymbolicExpression<Number, variables...>> store{nullptr};
-    };
+    struct SymbolicExpressionData;
   }  // namespace InputSpecBuilders
 
   namespace Internal
@@ -2709,40 +2696,6 @@ Core::IO::InputSpec Core::IO::InputSpecBuilders::input_field(
       });
   return spec;
 };
-
-
-template <typename Number, Core::Utils::CompileTimeString... variables>
-Core::IO::InputSpec Core::IO::InputSpecBuilders::symbolic_expression(
-    std::string name, SymbolicExpressionData<Number, variables...> data)
-{
-  auto store =
-      data.store ? data.store : in_container<Utils::SymbolicExpression<Number, variables...>>(name);
-  const std::type_info& stores_to = store.stores_to();
-
-  StoreFunction<std::string> convert_to_symbolic_expression{[store](Storage& out, std::string&& in)
-      {
-        try
-        {
-          Utils::SymbolicExpression<Number, variables...> expr(std::move(in));
-          return store(out, std::move(expr));
-        }
-        catch (const Core::Exception& e)
-        {
-          // Make a string from all variables.
-          std::stringstream message;
-          message << "could not be parsed as symbolic expression with variables: ";
-          ((message << std::quoted(std::string_view(variables)) << " "), ...);
-
-          return StoreStatus::fail(message.str());
-        }
-      },
-      stores_to};
-
-  return parameter<std::string>(name, {
-                                          .description = data.description,
-                                          .store = convert_to_symbolic_expression,
-                                      });
-}
 
 
 template <typename T>
